Add files via upload
Hi everyone, I haven't had time to mess around with making the title centred and making the page resizable. I will take a look at that when I've got more of the stats stuff done, but I know a few of you wanted to have the background code. For the time being, the method i have used to create the GUI are in the CreateBackground.java folder. You can see how I've implemented the code in the ShowStats.java file. Let me know if theres anything wrong with the code or if you can easily resize stuff :)